+++ cmd/operations/recomb.js
@@ -21,26 +21,51 @@
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
+// @see https://sharp.dimens.io/en/stable/api-operation/#recomb
+
 // Strict mode.
 'use strict'
 
-// Package modules.
-const sharp = require('sharp')
+// Local modules.
+const queue = require('../../lib/queue')
+
+// Configure.
+const options = {
+  matrix: {
+    desc: 'Recombination matrix',
+    type: 'string'
+  }
+}
+
+// Command builder.
+const builder = (yargs) => {
+  const optionNames = Object.keys(options)
+  return yargs
+    .strict()
+    .example('$0 recomb "0.3588 0.7044 0.1368 0.2990 0.5870 0.1140 0.2392 0.4696 0.0912"', 'The recomb will be applied to the output, in this case a sepia filter has been applied')
+    .epilog('For more information on available options, please visit https://sharp.dimens.io/en/stable/api-operation/#recomb')
+    .options(options)
+    .global(optionNames, false)
+    .group(optionNames, 'Command Options')
+}
+
+// Command handler.
+const handler = (args) => {
+  const matrix = args.matrix.split(' ').map((el) => parseFloat(el))
+
+  return queue.push([ 'recomb', (sharp) => {
+    return sharp.recomb([
+      matrix.slice(0, 3),
+      matrix.slice(3, 6),
+      matrix.slice(6, 9)
+    ])
+  }])
+}
 
 // Exports.
 module.exports = {
-  BAND: [ 'red', 'green', 'blue' ],
-  BOOL: Object.keys(sharp.bool),
-  COLOURSPACE: Object.keys(sharp.colourspace),
-  CONTAINER: [ 'fs', 'zip' ],
-  DEPTH: [ 'onepixel', 'onetile', 'one' ],
-  FIT: Object.keys(sharp.fit),
-  FORMAT: [ 'jpeg', 'jpg', 'png', 'raw', 'tiff', 'webp' ],
-  GRAVITY: Object.keys(sharp.gravity),
-  KERNEL: Object.keys(sharp.kernel),
-  LAYOUT: [ 'dz', 'google', 'zoomify' ],
-  POSITION: Object.keys(sharp.position),
-  STRATEGY: Object.keys(sharp.strategy),
-  TIFF_COMPRESSION: [ 'ccittfax4', 'deflate', 'jpeg', 'lzw', 'none' ],
-  TIFF_PREDICTOR: [ 'float', 'horizontal', 'none' ]
+  command: 'recomb <matrix>',
+  describe: 'Recomb the image with the specified matrix',
+  builder,
+  handler
 }