+++ cmd/resizing/extend.js
@@ -21,7 +21,7 @@
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-// @see https://sharp.pixelplumbing.com/en/stable/api-operation/#trim
+// @see https://sharp.dimens.io/en/stable/api-resize/#extend
 
 // Strict mode.
 'use strict'
@@ -31,9 +31,25 @@ const queue = require('../../lib/queue')
 
 // Configure.
 const options = {
-  tolerance: {
-    desc: 'The percentage similarity',
-    defaultDescription: '10',
+  background: {
+    defaultDescription: 'rgba(0, 0, 0, 1)',
+    desc: 'Background colour, parsed by the color module',
+    type: 'string'
+  },
+  bottom: {
+    desc: 'Offset from the bottom edge',
+    type: 'number'
+  },
+  left: {
+    desc: 'Offset from the left edge',
+    type: 'number'
+  },
+  right: {
+    desc: 'Offset from the right edge',
+    type: 'number'
+  },
+  top: {
+    desc: 'Offset from the top edge',
     type: 'number'
   }
 }
@@ -43,7 +59,8 @@ const builder = (yargs) => {
   const optionNames = Object.keys(options)
   return yargs
     .strict()
-    .epilog('For more information on available options, please visit https://sharp.pixelplumbing.com/en/stable/api-operation/#trim')
+    .example('$0 extend 10 20 10 10 rgba(0,0,0,0)', 'Add 10 transparent pixels to the top, left, and right edges, and 20 to the bottom edge')
+    .epilog('For more information on available options, please visit https://sharp.dimens.io/en/stable/api-resize/#extend')
     .options(options)
     .global(optionNames, false)
     .group(optionNames, 'Command Options')
@@ -51,13 +68,21 @@ const builder = (yargs) => {
 
 // Command handler.
 const handler = (args) => {
-  return queue.push([ 'trim', (sharp) => sharp.trim(args.tolerance) ])
+  return queue.push([ 'extend', (sharp) => {
+    return sharp.extend({
+      background: args.background,
+      bottom: args.bottom,
+      left: args.left,
+      right: args.right,
+      top: args.top
+    })
+  }])
 }
 
 // Exports.
 module.exports = {
-  command: 'trim [tolerance]',
-  describe: 'Trim "boring" pixels from all edges that contain values within a percentage similarity of the top-left pixel',
+  command: 'extend <top> <bottom> <left> <right>',
+  describe: 'Extends/pads the edges of the image with the provided background colour',
   builder,
   handler
 }