+++ test/cmd/resizing/trim.js
@@ -22,7 +22,7 @@
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-// @see https://sharp.pixelplumbing.com/en/stable/api-operation/#extract
+// @see https://sharp.dimens.io/en/stable/api-resize/#trim
 
 // Strict mode.
 'use strict'
@@ -33,48 +33,51 @@ const sinon = require('sinon')
 const Yargs = require('yargs')
 
 // Local modules.
-const extract = require('../../../cmd/operations/extract')
 const queue = require('../../../lib/queue')
 const sharp = require('../../mocks/sharp')
+const trim = require('../../../cmd/resizing/trim')
 
 // Test suite.
-describe('extract', () => {
-  const cli = (new Yargs()).command(extract)
+describe('trim', () => {
+  const cli = (new Yargs()).command(trim)
 
   // Reset.
   afterEach('queue', () => queue.splice(0))
   afterEach('sharp', sharp.prototype.reset)
 
-  describe('<top> <left> <width> <height>', () => {
-    // Default offset and dimensions.
-    const top = '10'
-    const left = '20'
-    const width = '30'
-    const height = '40'
-
+  describe('..', () => {
     // Run.
-    beforeEach((done) => cli.parse([ 'extract', top, left, width, height ], done))
+    beforeEach((done) => cli.parse([ 'trim' ], done))
 
     // Tests.
-    it('must set the top, left, width, and height flags', () => {
-      const args = cli.parsed.argv
-      expect(args).to.have.property('top', parseInt(args.top, 10))
-      expect(args).to.have.property('left', parseInt(args.left, 10))
-      expect(args).to.have.property('width', parseInt(args.width, 10))
-      expect(args).to.have.property('height', parseInt(args.height, 10))
-    })
     it('must update the pipeline', () => {
       expect(queue.pipeline).to.have.length(1)
-      expect(queue.pipeline).to.include('extract')
+      expect(queue.pipeline).to.include('trim')
     })
     it('must execute the pipeline', () => {
       const pipeline = queue.drain(sharp())
-      sinon.assert.calledWith(pipeline.extract, {
-        top: parseInt(top, 10),
-        left: parseInt(left, 10),
-        width: parseInt(width, 10),
-        height: parseInt(height, 10)
-      })
+      sinon.assert.called(pipeline.trim)
+    })
+  })
+
+  describe('[threshold]', () => {
+    // Default threshold.
+    const threshold = '10'
+
+    // Run.
+    beforeEach((done) => cli.parse([ 'trim', threshold ], done))
+
+    // Tests.
+    it('must set the threshold flag', () => {
+      expect(cli.parsed.argv).to.have.property('threshold', parseInt(threshold, 10))
+    })
+    it('must update the pipeline', () => {
+      expect(queue.pipeline).to.have.length(1)
+      expect(queue.pipeline).to.include('trim')
+    })
+    it('must execute the pipeline', () => {
+      const pipeline = queue.drain(sharp())
+      sinon.assert.calledWith(pipeline.trim, parseInt(threshold, 10))
     })
   })
 })